+++ ManiacEngine/ManiacEngine/DrawableObject.h
@@ -17,7 +17,7 @@ public:
 	DrawableObject();
 	//~DrawableObject();
 	virtual void Render(glm::mat4 globalModelTransform) = 0;
-	virtual void Update(float deltaTime);
+	virtual void Update(int deltaTime);
 	glm::vec3 GetPos(); 
 	glm::vec3 GetSize(); 
 	void SetSize(float sizeX, float sizeY);
